add test, fix implementation, remove extra subpaths, split components
