+++ packages/@react-spectrum/s2-ai/test/AssetList.test.tsx
@@ -1,0 +1,29 @@
+/*
+ * Copyright 2026 Adobe. All rights reserved.
+ * This file is licensed to you under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License. You may obtain a copy
+ * of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+ * OF ANY KIND, either express or implied. See the License for the specific language
+ * governing permissions and limitations under the License.
+ */
+
+import {Asset, AssetList} from '../src/AssetList';
+import {Image} from '@react-spectrum/s2/Image';
+import {render} from '@react-spectrum/test-utils-internal';
+
+describe('AssetList', () => {
+  it('should render', () => {
+    let {getByRole} = render(
+      <AssetList aria-label="Uploaded files">
+        <Asset aria-label="Demo file.pdf" textValue="Demo file.pdf">
+          <Image slot="thumbnail" src="https://example.com/image.png" />
+        </Asset>
+      </AssetList>
+    );
+
+    expect(getByRole('grid')).toBeInTheDocument();
+  });
+});
