+++ packages/@react-spectrum/s2/package.json
@@ -45,13 +45,6 @@
       "import": "./style/dist/module.mjs",
       "require": "./style/dist/main.cjs"
     },
-    "./style/runtime": {
-      "source": "./style/runtime.ts",
-      "types": "./dist/types/style/runtime.d.ts",
-      "module": "./style/dist/runtime.mjs",
-      "import": "./style/dist/runtime.mjs",
-      "require": "./style/dist/runtime.cjs"
-    },
     "./icons/*": {
       "source": "./s2wf-icons/S2_Icon_*_20_N.svg",
       "types": [
