+++ packages/react-aria/src/tag/useTag.ts
@@ -19,7 +19,6 @@ import intlMessages from '../../intl/tag/*.json';
 import {KeyboardEvent} from 'react';
 import type {ListState} from 'react-stately/useListState';
 import {mergeProps} from '../utils/mergeProps';
-// @ts-ignore
 import {SelectableItemStates} from '../selection/useSelectableItem';
 import {useDescription} from '../utils/useDescription';
 import {useFocusable} from '../interactions/useFocusable';
